Compilation works for 64bit of all variants

+++ jni/application/onscripter/backtrace.cpp
@@ -28,7 +28,7 @@ _Unwind_Reason_Code trace_func(struct _Unwind_Context *context, void* a)
     return _URC_NO_REASON;
 }
 
-unsigned int linux_gcc_demangler(const char *mangled_name, char *unmangled_name, size_t buffersize)
+size_t linux_gcc_demangler(const char *mangled_name, char *unmangled_name, size_t buffersize)
 {
 	size_t out_len = 0;
 	int status = 0;
@@ -44,7 +44,7 @@ unsigned int linux_gcc_demangler(const char *mangled_name, char *unmangled_name,
 	return out_len;
 }
 
-unsigned int get_backtrace(char** out, unsigned int buffSize)
+size_t get_backtrace(char** out, size_t buffSize)
 {
     const int max_frames = 100;
     const int name_max_length = 512;
@@ -55,7 +55,7 @@ unsigned int get_backtrace(char** out, unsigned int buffSize)
 
     char tmp[name_max_length];
     char line[name_max_length];
-    unsigned int offset;
+    size_t offset;
     size_t buf_length = 0;
     Dl_info info;
     void * ip;
@@ -69,24 +69,24 @@ unsigned int get_backtrace(char** out, unsigned int buffSize)
 
         // Get the data of this frame
         if (dladdr(ip, &info)) {
-            offset = (unsigned int)ip - (unsigned int)info.dli_saddr;
+            offset = (size_t)ip - (size_t)info.dli_saddr;
 
             const char* fn_name = info.dli_sname;
             const char* lib_name = info.dli_fname;
             if (linux_gcc_demangler(fn_name, tmp, name_max_length) != 0)
                 fn_name = tmp;
-            buf_length = sprintf(line, "\n    at %s.%x[%s]+%i", lib_name, (unsigned int)ip, fn_name, offset);
+            buf_length = sprintf(line, "\n    at %s.%x[%s]+%i", lib_name, (size_t)ip, fn_name, offset);
         } else {
-            buf_length = sprintf(line, "\n    <Unknown native code>.%x", (unsigned int)ip);
+            buf_length = sprintf(line, "\n    <Unknown native code>.%x", (size_t)ip);
         }
 
         // Check if we can copy the line in the buffer; +1 for terminating character
-        if (buf_length + ((unsigned int)out_buff - (unsigned int)*out) + 1 < buffSize) {
+        if (buf_length + ((size_t)out_buff - (size_t)*out) + 1 < buffSize) {
             strncpy(out_buff, line, buf_length);
             out_buff += buf_length;
         }
     }
-    unsigned int length = (unsigned int)out_buff - (unsigned int)*out;
+    size_t length = (size_t)out_buff - (size_t)*out;
     *(*out + length) = 0;
     return length;
 }
